refactor test and cleanup
