+++ db/db-test/db.test.js
@@ -1,18 +1,16 @@
 const dbHelpers = require('../dbHelpers');
 const mongoose = require('mongoose');
-const seed = require('./fakeData');
+const seed = require('./testData');
 
-describe('Database Tests', () => {
+describe('database helpers - save and find', () => {
   beforeAll((done) => {
     mongoose.connect('mongodb://localhost/testDB');
     const db = mongoose.connection;
     db.on('error', console.error.bind(console, 'connection error'));
     db.once('open', () => {
-      console.log('We are connected to test database!');
       seed.TestColl.remove({}, () => {
         dbHelpers.save(seed.testData, seed.TestColl, (result) => {
           if (result) {
-            console.log('====>', result);
             done();
           }
         });
@@ -20,22 +18,29 @@ describe('Database Tests', () => {
     });
   });
 
-  // Save object with 'name' value of 'Mike"
-  it('should save data the correct data to test database', async (done) => {
+  it('should save all data items to test database', (done) => {
     expect.assertions(2);
-    dbHelpers.find(90976, seed.TestColl, (err, docs) => {
+    dbHelpers.find({}, seed.TestColl, (err, docs) => {
       if (err) { throw err; }
-      console.log('data ===> ', docs);
-      expect(docs.length).toBe(1);
+      expect(docs.length).toBe(8);
       expect(docs[0].id).toBe(90976);
+      done();
+    });
+  });
+  // nb: the sample data contains a duplication of the restaurant with id: 89104
+  it('should not save duplicate restaurants', async (done) => {
+    expect.assertions(2);
+    dbHelpers.find({ id: 89104 }, seed.TestColl, (err, docs) => {
+      if (err) { throw err; }
+      expect(docs.length).toBe(1);
+      expect(docs[0].id).toBe(89104);
       done();
     });
   });
 
   afterAll((done) => {
     mongoose.connection.db.dropDatabase(() => {
-      mongoose.connection.close();
-      done();
+      mongoose.connection.close(done);
     });
   });
 });
