+++ package.json
@@ -4,7 +4,7 @@
   "description": "the menu component for our FEC project",
   "main": "src/server/server.js",
   "scripts": {
-    "test": "jest",
+    "test": "jest db.test.js",
     "build": "./node_modules/.bin/webpack -d --watch",
     "seed": "node db/dbHelpers.js"
   },
