+++ src/server/server.js
@@ -8,6 +8,7 @@ app.use(morgan('dev'));
 app.use('/', express.static(path.join(__dirname, '../client')));
 app.use('/', express.static(path.join(__dirname, '../public')));
 
+
 const port = 3005;
 
 app.listen(port, () => {});
